RpcTimeout从Long类型改成Int

+++ lealone-common/src/main/java/org/lealone/net/NetNodeManager.java
@@ -30,7 +30,7 @@ public interface NetNodeManager {
 
     String[] assignNodes(IDatabase db);
 
-    default long getRpcTimeout() {
+    default int getRpcTimeout() {
         return 0;
     }
